jobscraper/requirements.txt
# requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.14,<4.0
django-filter>=2.4,<25.0
django-cors-headers>=4.0,<5.0
beautifulsoup4>=4.12,<5.0
selenium>=4.10,<5.0
lxml>=4.9,<5.0
requests>=2.30,<3.0
feedparser>=6.0,<7.0  # <-- Add this line